All tests passed. Cannot get proper game play.
